feat: 졸업생용 권한 설정 및 클라이언트 페이지 분리 #217
